[setup] clean up and organise the repo
