+++ utilities/gpg_keys/quangvu3.asc
@@ -1,0 +1,51 @@
+-----BEGIN PGP PUBLIC KEY BLOCK-----
+
+mQINBF1WfAgBEACqpdhSxhXIRlsTSXxDUranvOKLa7cZM/oujhUO+/a8jcfVmwdr
+IhT1MgsogjXDuIDyIrVcGCl9IqsvJThbXO+4lYzoTMty+pILv+Lj4dvmJP4yyVTE
+r3sgtYYTHvqTuBgGq7V3K7KI1tyoqMQMLeG8zoZAgfJMBjPy4L2nDTrUkSWeLTlG
+LvjZ9XhF+2IitxSmoaYIV4oo6/8/xI+l+U1yRJZNfCLwwNZr225DVyTtxfzs95xE
+P6n8qMgztwo2sYmtUJchi/TerjBpJC8P/wpvg2X4q0v8XJQrtxkVavhmIWB7XV3K
+5EKj3aoLWyziacavfs3X+6h4tszM6+4O+Nbvg2fHnDR1BZxiXSbZA0s2l55ao/s1
+MaNOefXxtm9AQCXDTXC4dPdGI+H606aZ6aWYXbU5yWQnj6AafhD4rsPY2vm40wYI
+x7rJD8fw2vfOBaB8K0M9z2MJ6K/FK1i5RjuJwsa2qGlGU4Kf8B/miWyDGtful+oM
+QHasNDt/KfzJPU9wkowNZlZv7U154wCmOBH+W6CrJ9AhWEzId36ZIqwimlYTzfL5
+86PgoWTPhQjLpZ40rVI3nDsQmMHd+y1LuiS7CScVM92OHDDhMUaNrhX935lpAWtO
+rBkAjYyU86Ff3C26NhQi9nlp2j0JcctX14B+aac69tZTIZk8PAcFFqq8hQARAQAB
+tCBWdSBRdWFuZyA8amltbXl2dTIwMDhAZ21haWwuY29tPokCTgQTAQgAOBYhBJnx
+RP4eo/wEKersMCBApy/WJzUUBQJdVnwIAhsDBQsJCAcCBhUKCQgLAgQWAgMBAh4B
+AheAAAoJECBApy/WJzUUqpIP/3AcWtQLM+gyJdqy/b0HnFuVL1UolPHI3y8JXDN7
+b0bcXUGhInCkAdCUSgDCVUSOMsMM2AIRJTrMo0GO/Cd8FYQV0sCDLn62ppP3ropk
+bYPqOFUhqCoXxkSBP/bR5v+nJzx1MvwGwiLZ9NUhtN352B8MPOHdUk801VfAhhS5
+f3RBW3I4Lu6h68lKAspcK90INSXGPBz3imuuXupoK+YzeOB+yNiP8Cvvb8PGoB0T
+ZKgU0Uo14VNl4P084gNBjttYKcJhtvngYR/DODSrrL2qSma3Kp7gFPRLBsddd/Ie
+rqXHrDnXwhAJZUWHedBrJjCdK6+PMtaI6xRGAS5fw8hOa8mgoLk76Q5E+Bo17DQG
+z9naMWXL/OLuBE06o8DEpWZfBYnBd89NoTfr9CydkVcW8DlZSDXeAFW6LqOtLk9e
+7lDr6ZaW7PMsO17bMif90N7ZgnAN7hb0sFgp64+zuASBeQm2HIGvd6BGIL8fqrFc
+4BhfnVoXW0tu+IAVhFvKXMKck3wCmjC9Ouz2pBnHGHfSWe3qUJmGk9KVVx0plvKD
+sXvHPVfvMms7b03ssDoawxuqD1jIj1OcMDLbuUJpcDUlUITbylrRH+3984QTsn8o
+vathGpX0fhEL5U7GSoIkXSQD0MFCnoZ4VThnpAN9GccO9Su2d6EduAqdQmyfYzkJ
+m0W4uQINBF1WfAgBEACycfO40xELIOwJKubzuoH06Ip0h7O2DqzHsvd2ws+LyIWU
+7WYaP61KEfX2AzIuvWc/dyoCb31bvKNb1/bpb74IU8BMu3wcfAVoIFx8DHU+PntC
+R+kQXOC8IDNjOyyoiDxl5sM6GTTzjz7Cy+1NmZzbwdbChLolngPrkyFnRq9UZdmn
+1LEHRwMcCd8GwsonNfkHdzlBAvZKfU7Pu0ffCV3xyrcwf9L40MM8+ms3mAdw50eb
+wZmVaghCYLh6yT7Exw85h+1dnBeArKh+CojHY4ChuoiLPeGrburbkqLsOCM4U5pN
+zAJPLCL3155srHj8LwxdPDgmzTZbdxNW39aQLCaZ+3lbcXWp3FdHB7RS3sIRpzQR
+pFt+yikRhKsESh6uf22dtE+MjnqrqGQZwjMcUh0uokgZ8qmzYx+S3LPLLZ3ACwlO
+JDGZJe3TfFAhy4LY6/I+qk0Ryhq1TfBDZ83ZMQlQgiwiqmQZnlXBI1m7HTSUwPEa
+q0vDZWQdYX65s1jpIjpFTklsCUXBcOC15zpj/mA2msJ9TimZf+ziCDJL3jD+LTxW
+b+F6drjS8ROeZUh9itMjp8QiFrbEdWs6WJm/SpJPXlHeqoFx6YvWL/UyEhsKFwKf
+MF1zAO858R6awgF7ndOXkvNLJePJ43J0prWL9WAN4kIH0Q9CEggUq6xcDdpWrQAR
+AQABiQI2BBgBCAAgFiEEmfFE/h6j/AQp6uwwIECnL9YnNRQFAl1WfAgCGwwACgkQ
+IECnL9YnNRTxWA/+NdV+hCUEjdZX+evTLxmV2xjJTo1gMPedEd6h/GLRkBUOghTr
+EnWQX6ezwZiEVF9SeuM65bfLjhHRBoVFt32F+MrBsnuZxITwgmLoDv5x8rKo+tH1
+2KlqS8ZRapyWXZu/5gNfikQlPy20TIW1npMN0JkqWaNlK2VFQJ01koPX+GHc8mvq
+ulllCToFrg/49xyeOth+rvWoqrxZXi3oKHs0wCga0y1iThKNC+1GBWi8QWB/3G8W
+f1bPPqIbGwnspO21XUNARE7u37tFmNtqzgd284wSQXjMUarJsDVbNkJltTicCjkv
+3tWtnLZRABlZai6zUssGJpbgL+pxOv8yFvZyjQ80siDYCV3o77kHBmKNZwcdVdm7
+AF84l60KSulvYv8w84r+H0iZiywXO46NdyK3dhuzHVWVLdVj7g92/epHqUb9u/Yl
+SR+mPLKZbvhODkVQgBb7Y5aCaKbOJBeDfVVKfrXrVp7F/a8oBmR5k1TxHEi7+bRY
+oEBGooU8+2h7weM6gcs9gX+23ngUpR1VZTLTWqudW0QC8WsYewiRNYz4vCWFOIDS
+eX5tg6t4tV56kUlIfdBZfey36bmISk9PAon8CUfurV4wRRvEDukTXvqbD/wn9DfP
+IXKZB3IfsDRcpm29yoRCkhGRK+NFrflt3+vozCz0GiibLOByz+zlNhYg7Ns=
+=Hr5B
+-----END PGP PUBLIC KEY BLOCK-----
