+++ utilities/munin_plugins/outgoing_connections_count
