--- utils/gpg_keys/sumogr.asc
@@ -1,30 +1,0 @@
------BEGIN PGP PUBLIC KEY BLOCK-----
-
-mQENBF3a3l8BCADdegpPeeMxByVPWoXMRfr48z2rj28xcLKS2ikiPJMrgUVjjckx
-uDXSqZL3nSIRDwUbRNEgJ+PJNw2/FESSyckM4xm5VOJmOxkVkl9vUlwGQvoZaR2Q
-sBhlRPe0iRrKqmLVeK4cfpE4qyB+8tOm0iISpVetqut6tMjl1lEz/kaiGZD6y2eO
-iiSUF+N1fai/DHSc+h/23cuUyKBOMFudKBb6Gs7KcBNzcfGWXwwvOOk8CtQwX6J7
-m/h4cuarniEJTdQOmGDEWKx6GN5q10vDXD6pU3MGH2o+Zze6dNd/LTtTXAsAkhZr
-ncbGkJvU+HwpIQwJpuZzBeaoD4cY6z6FfSrzABEBAAG0GHN1bW9nciA8c3Vtb2dy
-QG1haWwuY29tPokBVAQTAQgAPgIbAwULCQgHAgYVCgkICwIEFgIDAQIeAQIXgBYh
-BOz4gO+MRuNjSlPEqjhbU/PiqU6FBQJd2t9cBQkWkdTBAAoJEDhbU/PiqU6F9mwH
-/17Wts7aUyI4GsK98ByGzjQQz8MH2MDY2OwE0tHHaVaAaRWAVcZvbp0sQVEND/3W
-w/8Ai7OZz3eQAJUXqG2wm3Gvh6B7GFxuHeltnyDNNN16qUoxYv90kQ65dFvtA3fD
-SLesJnT9jqGwvpQLGSjYGTTVWvhb3RP2TyJJ8nCNo3WRpq9GTNUe25kglkSLM1St
-ItKyPNGQ9YD9cs2ucYfmD4GHpBz7qU/ruuLrEGB8Qmdi4gXrlp8xmIiNEyPSGxUz
-853ZJTMNlbt+IyVekb3owQSmcuyUDvOW69CTJT0S0MRjW/sZd6lmMmhRDYP/LRgx
-RCUXOHgSwLnFrdoFkDr2HaK5AQ0EXdreXwEIALAO6bHW9bXFo2ajuwN1EPt+j3AB
-77XPB5lA0H1RCsZfaNNjd/Sw3y1ykJu1LloUBT2H22c++Tdg8ChojVAPPJO7ILXf
-wYqGhzS8cjetaGvNfPaLRoY6xQgv9lNO9lcwIcGCH77AKgbyu0EeCaKiq4HtetZy
-XKVgt2nBXq7aFypqy43atmIYwSpSGrcdUZy34C38/lhRxxuIsZtrOWushQfVaARm
-CM97o4/5wyrhvLV9FOTFDAG0iS6ztVtRK7Oe+LB6HULmqMXxpxYy3/Ev52mNNC2s
-o283eFjda0evkb0uqyAl1BaozqKP2Q96mhhxvM1kNXNWqNFl9dx9lUHh5q8AEQEA
-AYkBPAQYAQgAJhYhBOz4gO+MRuNjSlPEqjhbU/PiqU6FBQJd2t5fAhsMBQkDwy7B
-AAoJEDhbU/PiqU6FoPcIAJv5IxbrDjilCOL4CsK/fPOaAn4T/vUQpkFerVhyGJsb
-QkffjGIr3mvonMnkBcVz8ODmcS4mfz/V8JJz9rkmkJvZ4OSdJ9QKHVNQkoGc2ugY
-oHdOweAh6kF2ZS1RFzve5c9DlEikWQK4AOKgv+Zjbb2maCGHX0ROqCRjcX5VAeSg
-l03iSx3NJtCIFxv9ntkMGJ8hWGYCeFVLxEoK8rOcZCJzzg29mkOAZmU8eXV1Si50
-iihsCyyZMgbZFwSV87aOEGVROBi5SNIQE1o/on7wv7RGO2sZv5tyJ1OJVwHx9eKR
-IFKPL52Gr0TgVYI7h2da2IOpO0theB6ZhyrLZGRmhDU=
-=Tn4w
------END PGP PUBLIC KEY BLOCK-----
